--- components/atoms/Avatar.tsx
@@ -1,6 +1,0 @@
-const Avatar = ({name,src}) => {
-    return ( <>
-    </> );
-}
- 
-export default Avatar;
